requirements/tests.txt
ipdb
pytest==4.3.0
flake8==3.5.0
pytest-cov==2.6.1
pytest-pythonpath==0.7.3
vcrpy==2.0.1
pytest-vcr==1.0.2
pytest-django~=3.4.0
pytest-faker~=2.0.0
pytest-factoryboy==2.0.2
freezegun
